backend_v2/requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
pydantic==2.12.3
python-dotenv==1.2.1
aiohttp==3.13.2
pandas==2.3.3
openpyxl==3.1.5
pyyaml==6.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
